pyproject.toml
[tool.poetry]
authors = ["KnowYourselves <rsalvarez@uc.cl>"]
description = "Rentool Backend"
name = "rentool-backend"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^3.2.4"
Pillow = "^9.1.0"
dj-database-url = "^0.5.0"
django-cors-headers = "^3.11.0"
django-extensions = "^3.1.5"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "4.3.0"
djoser = "^2.1.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
python = "^3.9"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
PyJWT = "1.7.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[tool.isort]
ensure_newline_before_comments = true
extend_skip = ["migrations", "manage.py"]
filter_files = true
force_single_line = true
group_by_package = true
import_heading_django = "django"
import_heading_stdlib = "standard library"
import_heading_thirdparty = "others libraries"
known_django = "django"
lines_between_types = 1
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
